# backend/docs/tesfafund-openapi.yaml
...
openapi: 3.0.0
info:
  version: 0.1.0
  title: TesfaFund API
  description: API specification for the TesfaFund platform.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
servers:
  - url: http://localhost:4000
...
paths:
...
  /campaigns:
    # DONE
    # (Role: Supervisor, Owner: No) => Full access
    # (Role: Recipient, Owner: Yes) => Full access
    # (Role: Anon | Recipient, Owner: No) => Access to public campaigns with ensitive fields hidden
    get:
      tags:
        - Campaigns
      summary: Retrieves a list of a campaigns, with optional filtering and pagination based on user's authorization
      parameters:
        - name: title
          description: Campaign title to search by.
          in: query
          schema:
            type: string
        - name: status
          description: Filter by campaign status.
          in: query
          schema:
            type: string
            enum: [PENDING_REVIEW, VERIFIED, DENIED, LIVE, PAUSED, COMPLETED]
        - name: category
          description: Filter by campaign category.
          in: query
          schema:
            type: string
        - name: minLaunchDate
          description: Minimum launch date of campaign (inclusive).
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: maxLaunchDate
          description: Maximum launch date of campaign (inclusive).
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: minSubmissionDate
          description: Minimum submission date of campaign (inclusive). This option is available only for Supervisors and Recipients.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: maxSubmissionDate
          description: Maximum submission date of campaign (inclusive). This option is available only for Supervisors and Recipients.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: minVerificationDate
          description: Filter by minimum campaign verification date. This option is available only for Supervisors and Recipients.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: maxVerificationDate
          description: Filter by maximum campaign verification date. This option is available only for Supervisors and Recipients.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: minDenialDate
          description: Earliest denial date of campaign application (inclusive). This option is available only for Supervisors and Recipients.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: maxDenialDate
          description: Latest denial date of campaign application (inclusive). This option is available only for Supervisors and Recipients.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: minEndDate
          description: Filter by minimum campaign completion date.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: maxEndDate
          description: Filter by maximum campaign completion date.
          in: query
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: isPublic
          description: Filters public campaigns.
          in: query
          schema:
            type: boolean
        - name: ownerRecipientId
          description: The ID of the recipient that owns the campaign
          in: query
          schema:
            type: string
            format: uuid
        - name: page
          description: Page number for pagination. Must be a positive integer.
          in: query
          schema:
            type: integer
        - name: limit
          description: Maximum number of items per page. Must be a positive integer.
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Successful retrieval
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedList"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/Campaign"
        "400":
          description: Bad request - Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: No campaigns found matching the criteria (also used instead of 403 forbidden).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
    # DONE
    # (Role: Supervisor, Owner: No) => No access
    # (Role: Recipient, Owner: Yes | No) => Full access
    # (Role: Anon, Owner: No) => No access
    post:
      tags:
        - Campaigns
      summary: Creates a campaign application for the authenticated Recipient.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              allOf:
                - $ref: "#/components/schemas/Campaign"
                - type: object
                  properties:
                    campaignDocuments:
                      type: array
                      items:
                        type: string
                        format: binary
                      description: Array of campaign document file names to upload.
                  required:
                    - campaignDocuments
      responses:
        "201":
          description: Campaign application submitted successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "400":
          description: Bad request - Invalid request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized - Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
  # Permissions differ by method.
  /campaigns/{id}:
    # DONE
    # (Role: Supervisor, Owner: No) => Full access
    # (Role: Recipient, Owner: Yes) => Full access
    # (Role: Anon | Recipient, Owner: No) => Access to public campaigns with ensitive fields hidden
    get:
      tags:
        - Campaigns
      summary: Retrieves a campaign by ID, based on user's authorization
      description: |
        Retrieves a campaign by ID, respecting requester authorization: public campaign (unauthenticated), Recipient's campaign (authenticated), or all campaigns (Supervisor).
      parameters:
        - name: id
          in: path
          required: true
          description: The unique id of the campaign.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful retrieval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "400":
          description: Bad request - Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized - Authentication required or invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Campaign with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
    # DONE
    # (Role: Supervisor, Owner: No) => Full access
    # (Role: Recipient, Owner: Yes) => No access
    # (Role: Anon | Recipient, Owner: No) => No access
    put:
      tags:
        - Campaigns
      summary: Updates the campaign data. Available only for supervisors.
      description: This is used exclusively by supervisors to update a campaign (without recipient involvement), or when they approve the campaign application.
      parameters:
        - name: id
          in: path
          required: true
          description: The unique id of the campaign to update.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Campaign"
      responses:
        "204":
          description: Successful update - No content to return.
        "400":
          description: Bad request - Invalid request body or parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Recipient with given ID not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
  # Permissions differ by method
...
  /campaign-requests:
    # NEXT 10
    # (Role: Supervisor, Owner: No) => Full access
    # (Role: Recipient, Owner: Yes) => Full access
    # (Role: Anon | Recipient, Owner: No) => No access
    get:
      tags:
        - Campaign Requests
      summary: Retrieves a list of campaign requests, with optional filtering and pagination.
      description: |
        Retrieves campaign requests, granting access based on Recipient or Supervisor authorization. Provides optional filtering and pagination.
      parameters:
        - name: campaignId
          in: query
          description: The ID of the campaign.
          schema:
            type: string
        - name: requestType
          in: query
          description: Filter by request type.
          schema:
            type: string
            enum:
              [
                GOAL_ADJUSTMENT_REQUEST,
                STATUS_CHANGE_REQUEST,
                POST_UPDATE_REQUEST,
                END_EXTENSION_REQUEST,
              ]
        - name: isResolved
          in: query
          description: Filter by resolution status (true/false).
          schema:
            type: boolean
        - name: minRequestDate
          in: query
          description: Filter requests created after this date (inclusive).
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: maxRequestDate
          in: query
          description: Filter requests created before this date (inclusive).
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: minResolutionDate
          in: query
          description: Filter requests resolved after this date (inclusive).
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: maxResolutionDate
          in: query
          description: Filter requests resolved before this date (inclusive).
          schema:
            oneOf:
              - type: string
                format: date-time
              - type: string
                format: date
        - name: title
          description: Search for requests with this title
          in: query
          schema:
            type: string
        - name: page
          in: query
          description: Page number for pagination.
          schema:
            type: integer
        - name: limit
          in: query
          description: Maximum number of items per page.
          schema:
            type: integer
      responses:
        "200":
          description: Successful retrieval.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedList"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          oneOf:
                            - $ref: "#/components/schemas/GoalAdjustmentRequest"
                            - $ref: "#/components/schemas/StatusChangeRequest"
                            - $ref: "#/components/schemas/PostUpdateRequest"
                            - $ref: "#/components/schemas/EndDateExtensionRequest"
        "400":
          description: Bad request - Invalid parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized - Authentication required or invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Campaign or adjustment requests not found (also used instead of 403 forbidden).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
    # NEXT 9
    # (Role: Supervisor, Owner: No) => No access
    # (Role: Recipient, Owner: Yes) => Full access
    # (Role: Anon | Recipient, Owner: No) => No access
    post:
      tags:
        - Campaign Requests
      summary: Creates campaign requests for the authenticated Recipient's campaign.
      parameters:
        - name: campaignId
          in: query
          required: true
          description: The ID of the campaign.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/GoalAdjustmentRequest"
                - $ref: "#/components/schemas/StatusChangeRequest"
                - $ref: "#/components/schemas/PostUpdateRequest"
                - $ref: "#/components/schemas/EndDateExtensionRequest"
      responses:
        "201":
          description: Successful request submission.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/GoalAdjustmentRequest"
                  - $ref: "#/components/schemas/StatusChangeRequest"
                  - $ref: "#/components/schemas/PostUpdateRequest"
                  - $ref: "#/components/schemas/EndDateExtensionRequest"
        "409":
          description: Conflict - Request already exists, or resource in incorrect state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "400":
          description: Bad request - Invalid request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized - Authentication required or invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
  # Permissions differ by method
  /campaign-requests/{id}:
    # NEXT 11
    # (Role: Supervisor, Owner: No) => Full access
    # (Role: Recipient, Owner: Yes) => Full access
    # (Role: Anon | Recipient, Owner: No) => No access
    get:
      tags:
        - Campaign Requests
      summary: Retrieves a campaign request by ID. Access depends on user's role.
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the campaign request
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful retrieval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignRequest"
        "400":
          description: Bad request - Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized - Authentication required or invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Campaign request with given Id not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
    # NEXT 13
    # (Role: Supervisor, Owner: No) => Full access
    # (Role: Recipient, Owner: Yes) => Full access
    # (Role: Anon | Recipient, Owner: No) => No access
    delete:
      tags:
        - Campaign Request
      summary: Deletes a campaign request. Access depends on user's role.
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the campaign request
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Successful deletion
        "400":
          description: Bad request - Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized - Authentication required or invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Campaign request with given ID not found (also used inplace of 403 forbidden).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
      security:
        - bearerHttpAuth: []
  # Permissions differ by method
...
components:
  schemas:
...
    PaginatedList:
      type: object
      properties:
        page-number:
          type: integer
        page-count:
          type: integer
      required:
        - page-number
        - page-count
...
    Campaign:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        fundraisingGoal:
          type: string
        status:
          $ref: "#/components/schemas/CampaignStatus"
        category:
          type: string
        paymentInfo:
          $ref: "#/components/schemas/PaymentInfo"
        submissionDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
        verificationDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
          nullable: true
        denialDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
          nullable: true
        launchDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
          nullable: true
        endDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
          nullable: true
        documentUrls:
          type: array
          items:
            type: string
        redactedDocumentUrls:
          type: array
          items:
            type: string
      required:
        - title
        - description
        - fundraisingGoal
        - paymentInfo
        - category
        - endDate
    CampaignStatus:
      type: string
      enum:
        - PENDING_REVIEW
        - VERIFIED
        - DENIED
        - LIVE
        - PAUSED
        - COMPLETED
    CampaignPost:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        publicPostDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
        campaignId:
          type: string
          format: uuid
      required:
        - title
        - content
        - campaignId
    CampaignRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        requestDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
        justification:
          type: string
        resolutionDate:
          oneOf:
            - type: string
              format: date-time
            - type: string
              format: date
        campaignId:
          type: string
          format: uuid
      required:
        - title
        - justification
    GoalAdjustmentRequest:
      allOf:
        - $ref: "#/components/schemas/CampaignRequest"
        - type: object
          properties:
            newGoal:
              type: string
          required:
            - newGoal
    StatusChangeRequest:
      allOf:
        - $ref: "#/components/schemas/CampaignRequest"
        - type: object
          properties:
            newStatus:
              $ref: "#/components/schemas/CampaignStatus"
          required:
            - newStatus
    PostUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/CampaignRequest"
        - type: object
          properties:
            newPost:
              $ref: "#/components/schemas/CampaignPost"
          required:
            - newPost
    EndDateExtensionRequest:
      allOf:
        - $ref: "#/components/schemas/CampaignRequest"
        - type: object
          properties:
            newEndDate:
              oneOf:
                - type: string
                  format: date-time
                - type: string
                  format: date
          required:
            - newEndDate
    PaymentInfo:
      type: object
      properties:
        paymentMethod:
          type: string
          description: The payment method used.
        phoneNo:
          type: string
          description: The phone number associated with the payment.
        bankAccountNo:
          type: string
          description: The bank account number for bank transfers.
        bankName:
          type: string
          description: The name of the bank.
      required:
        - paymentMethod
      anyOf:
        - required:
            - phoneNo
        - required:
            - bankAccountNo
            - bankName
...
    ProblemDetails:
      type: object
      properties:
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
      additionalProperties: true
  securitySchemes:
    bearerHttpAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token using a JWT